requirements.txt
gitdb2==3.0.2
GitPython==3.0.8
smmap2==2.0.5
